chore: 코드 스타일 수정 및 파일 위치 조정
